+++ Kino.Infrastructure/Services/MovieCrewService.cs
@@ -9,10 +9,12 @@ namespace Kino.Infrastructure.Services
     public class MovieCrewService : IMovieCrewService
     {
         private readonly IMovieCrewRepository _movieCrewRepository;
+        private readonly IDepartmentRepository _departmentRepository;
 
-        public MovieCrewService(IMovieCrewRepository movieCrewRepository)
+        public MovieCrewService(IMovieCrewRepository movieCrewRepository, IDepartmentRepository departmentRepository)
         {
             _movieCrewRepository = movieCrewRepository;
+            _departmentRepository = departmentRepository;
         }
 
         public async Task<IEnumerable<MovieCrewDetailResponse>?> GetMovieCrewsDetailsByMovieId(int id)
@@ -25,28 +27,33 @@ namespace Kino.Infrastructure.Services
                 MovieId = x.MovieId,
                 PersonId = x.PersonId,
                 PersonName = x.Person.PersonName,
-                DepartmentId = x.DepartmentId,
-                DepartmentName = x.Department.DepartmentName
+                Department = x.Department.DepartmentName
             });
             return response;
         }
 
         public async Task<bool> AddMovieCrew(MovieCrewRequest movieCrewRequest)
         {
+            var department = await _departmentRepository.SingleOrDefaultAsync(x => x.DepartmentName == movieCrewRequest.Department);
+            if (department == null) 
+                return false;
             var crew = new MovieCrew
             {
                 MovieId = movieCrewRequest.MovieId,
                 PersonId = movieCrewRequest.PersonId,
-                DepartmentId = movieCrewRequest.DepartmentId
+                DepartmentId = department.Id
             };
             return await _movieCrewRepository.AddAsync(crew);
         }
 
         public async Task<bool> DeleteMovieCrew(MovieCrewRequest movieCrewRequest)
         {
+            var department = await _departmentRepository.SingleOrDefaultAsync(x => x.DepartmentName == movieCrewRequest.Department);
+            if (department == null)
+                return false;
             var crew = await _movieCrewRepository.SingleOrDefaultAsync(x => x.MovieId == movieCrewRequest.MovieId
                                                                             && x.PersonId == movieCrewRequest.PersonId
-                                                                            && x.DepartmentId == movieCrewRequest.DepartmentId);
+                                                                            && x.DepartmentId == department.Id);
             if (crew == null)
                 return false;
             return await _movieCrewRepository.RemoveAsync(crew);
@@ -54,9 +61,12 @@ namespace Kino.Infrastructure.Services
 
         public async Task<bool> MovieCrewExists(MovieCrewRequest movieCrewRequest)
         {
+            var department = await _departmentRepository.SingleOrDefaultAsync(x => x.DepartmentName == movieCrewRequest.Department);
+            if (department == null)
+                return false;
             return await _movieCrewRepository.AnyAsync(x => x.MovieId == movieCrewRequest.MovieId
                                                             && x.PersonId == movieCrewRequest.PersonId
-                                                            && x.DepartmentId == movieCrewRequest.DepartmentId);
+                                                            && x.DepartmentId == department.Id);
         }
     }
 }
